% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kumojars-package.R
\docType{package}
\name{kumojars}
\alias{kumojars}
\alias{kumojars-package}
\title{Java Archive Wrapper Supporting the 'kumo' Package}
\description{
Java Archive Wrapper Supporting the 'kumo' Package
}
\author{
Bob Rudis (bob@rud.is)
}
